--- src/Chirp.Web/Pages/Private/SendCheep.cshtml
@@ -1,18 +1,0 @@
-﻿@page "/cheep/post"
-@model Chirp.Web.Pages.SendCheepModel
-
-@* https://stackoverflow.com/a/67320817/11014922 *@
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-
-@{
-	ViewData["Title"] = "Chirp!";
-	Layout = "Shared/_Layout";
-}
-
-<div class="cheepbox">
-	<h3>What's on your mind @(User.Identity.Name)?</h3>
-	<form method="post">
-		<input style="float: left" type="text" asp-for="CheepMessage">
-		<input type="submit" value="Share">
-	</form>
-</div>
