+++ src/Chirp.Web/Pages/Public/Public.cshtml
@@ -1,5 +1,6 @@
 ﻿@page "/"
 @model Chirp.Web.Pages.PublicModel
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @{
 	ViewData["Title"] = "Chirp!";
 	Layout = "Shared/_Layout";
@@ -7,6 +8,17 @@
 
 <div>
 	<h2> Public Timeline </h2>
+
+	@if (User.Identity?.IsAuthenticated == true)
+	{
+		<div class="cheepbox">
+			<h3>What's on your mind @(User.Identity.Name)?</h3>
+			<form method="post">
+				<input style="float: left" type="text" asp-for="CheepMessage">
+				<input type="submit" value="Share">
+			</form>
+		</div>
+	}
 
 	@if (Model.Cheeps.Any())
 	{
