+++ src/Chirp.Web/Pages/Public/UserTimeline.cshtml
@@ -1,33 +1,45 @@
 ﻿@page "/{author}"
 @model Chirp.Web.Pages.UserTimelineModel
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @{
-    ViewData["Title"] = "Chirp!";
-    Layout = "Shared/_Layout";
-    var routeName = HttpContext.GetRouteValue("author");
+	ViewData["Title"] = "Chirp!";
+	Layout = "Shared/_Layout";
+	var routeName = HttpContext.GetRouteValue("author");
 }
 
 <div>
-    <h2> @routeName's Timeline </h2>
+	<h2> @routeName's Timeline </h2>
 
-    @if (Model.Cheeps.Any())
-    {
-        <ul id="messagelist" class="cheeps">
-            @foreach (var cheep in Model.Cheeps)
-            {
-                <li>
-                    <p>
-                        <strong>
-                            <a href="/@cheep.AuthorName">@cheep.AuthorName</a>
-                        </strong>
-                        @cheep.Message
-                        <small>&mdash; @cheep.TimeStamp</small>
-                    </p>
-                </li>
-            }
-        </ul>
-    }
-    else
-    {
-        <em>There are no cheeps so far.</em>
-    }
+	@if (User.Identity?.IsAuthenticated == true)
+	{
+		<div class="cheepbox">
+			<h3>What's on your mind @(User.Identity.Name)?</h3>
+			<form method="post">
+				<input style="float: left" type="text" asp-for="CheepMessage">
+				<input type="submit" value="Share">
+			</form>
+		</div>
+	}
+
+	@if (Model.Cheeps.Any())
+	{
+		<ul id="messagelist" class="cheeps">
+			@foreach (var cheep in Model.Cheeps)
+			{
+				<li>
+					<p>
+						<strong>
+							<a href="/@cheep.AuthorName">@cheep.AuthorName</a>
+						</strong>
+						@cheep.Message
+						<small>&mdash; @cheep.TimeStamp</small>
+					</p>
+				</li>
+			}
+		</ul>
+	}
+	else
+	{
+		<em>There are no cheeps so far.</em>
+	}
 </div>
