+++ src/Chirp.Web/Pages/Public/UserTimeline.cshtml.cs
@@ -6,17 +6,58 @@ namespace Chirp.Web.Pages;
 
 public class UserTimelineModel : PageModel
 {
-	private readonly ICheepRepository _repository;
+	private readonly IAuthorRepository AuthorRepository;
+	private readonly ICheepRepository CheepRepository;
 	public required List<CheepDTO> Cheeps { get; set; }
 
-	public UserTimelineModel(ICheepRepository repository)
+
+	[BindProperty]
+	public string? CheepMessage { get; set; }
+
+	public UserTimelineModel(IAuthorRepository authorRepository, ICheepRepository cheepRepository)
 	{
-		_repository = repository;
+		AuthorRepository = authorRepository;
+		CheepRepository = cheepRepository;
 	}
 
 	public ActionResult OnGet(string author, [FromQuery(Name = "page")] int page = 1)
 	{
-		Cheeps = _repository.GetCheepsFromAuthor(page, author);
+		Cheeps = CheepRepository.GetCheepsFromAuthor(page, author);
 		return Page();
+	}
+
+	public IActionResult OnPost()
+	{
+		Console.WriteLine("OnPost called!");
+
+		string email = User.Claims.Where(a => a.Type == "emails").Select(e => e.Value).Single();
+
+		Console.WriteLine(email + " wrote: " + CheepMessage);
+
+		// check if author exists
+		// fetch author
+		// possibly create author
+
+		AuthorDTO? Author = AuthorRepository.GetAuthorByEmail(email).First();
+
+		if (Author == null)
+		{
+			Author = CreateNewAuthor();
+		}
+
+		CreateCheepDTO cheep = new CreateCheepDTO()
+		{
+			CheepGuid = new Guid(),
+			Text = CheepMessage,
+			Name = Author.Name,
+			Email = Author.Email
+		};
+
+		return Redirect("/");
+	}
+
+	private AuthorDTO CreateNewAuthor()
+	{
+		throw new NotImplementedException();
 	}
 }
